Refactored code into a utils file
